+++ examples/hybrid/schur_complement_W.jl
@@ -6,18 +6,6 @@ using ClimaCore.Utilities: half
 const compose = Operators.ComposeStencils()
 const apply = Operators.ApplyStencil()
 
-# Note: We denote energy variables with 𝔼, momentum variables with 𝕄, and tracer
-# variables with 𝕋.
-is_energy_var(symbol) = symbol in (:ρθ, :ρe_tot, :ρe_int)
-is_momentum_var(symbol) = symbol in (:uₕ, :ρuₕ, :w, :ρw)
-is_edmf_var(symbol) = symbol in (:turbconv,)
-is_tracer_var(symbol) = !(
-    symbol == :ρ ||
-    is_energy_var(symbol) ||
-    is_momentum_var(symbol) ||
-    is_edmf_var(symbol)
-)
-
 function FieldFromNamedTuple(space, nt::NamedTuple)
     cmv(z) = nt
     return cmv.(Fields.coordinate_field(space))
@@ -62,9 +50,9 @@ end
 
 function SchurComplementW(Y, transform, flags, test = false)
     @assert length(filter(isequal(:ρ), propertynames(Y.c))) == 1
-    @assert length(filter(is_energy_var, propertynames(Y.c))) == 1
-    @assert length(filter(is_momentum_var, propertynames(Y.c))) == 1
-    @assert length(filter(is_momentum_var, propertynames(Y.f))) == 1
+    @assert length(filter(CA.is_energy_var, propertynames(Y.c))) == 1
+    @assert length(filter(CA.is_momentum_var, propertynames(Y.c))) == 1
+    @assert length(filter(CA.is_momentum_var, propertynames(Y.f))) == 1
 
     FT = eltype(Y)
     dtγ_ref = Ref(zero(FT))
@@ -81,7 +69,7 @@ function SchurComplementW(Y, transform, flags, test = false)
     ∂ᶠ𝕄ₜ∂ᶜ𝔼 = Fields.Field(bidiag_type, axes(Y.f))
     ∂ᶠ𝕄ₜ∂ᶜρ = Fields.Field(bidiag_type, axes(Y.f))
     ∂ᶠ𝕄ₜ∂ᶠ𝕄 = Fields.Field(tridiag_type, axes(Y.f))
-    ᶜ𝕋_names = filter(is_tracer_var, propertynames(Y.c))
+    ᶜ𝕋_names = filter(CA.is_tracer_var, propertynames(Y.c))
     ∂ᶜ𝕋ₜ∂ᶠ𝕄_field =
         FieldFromNamedTuple(axes(Y.c), tracer_variables(FT, ᶜ𝕋_names))
 
@@ -259,13 +247,13 @@ function _linsolve_serial!(
 
     xᶜρ = xc.ρ
     bᶜρ = bc.ρ
-    ᶜ𝔼_name = filter(is_energy_var, propertynames(xc))[1]
+    ᶜ𝔼_name = filter(CA.is_energy_var, propertynames(xc))[1]
     xᶜ𝔼 = getproperty(xc, ᶜ𝔼_name)
     bᶜ𝔼 = getproperty(bc, ᶜ𝔼_name)
-    ᶜ𝕄_name = filter(is_momentum_var, propertynames(xc))[1]
+    ᶜ𝕄_name = filter(CA.is_momentum_var, propertynames(xc))[1]
     xᶜ𝕄 = getproperty(xc, ᶜ𝕄_name)
     bᶜ𝕄 = getproperty(bc, ᶜ𝕄_name)
-    ᶠ𝕄_name = filter(is_momentum_var, propertynames(xf))[1]
+    ᶠ𝕄_name = filter(CA.is_momentum_var, propertynames(xf))[1]
     xᶠ𝕄 = getproperty(xf, ᶠ𝕄_name).components.data.:1
     bᶠ𝕄 = getproperty(bf, ᶠ𝕄_name).components.data.:1
 
@@ -282,18 +270,18 @@ function _linsolve_serial!(
     @. xᶜρ = -bᶜρ + dtγ * apply(∂ᶜρₜ∂ᶠ𝕄, xᶠ𝕄)
     @. xᶜ𝔼 = -bᶜ𝔼 + dtγ * apply(∂ᶜ𝔼ₜ∂ᶠ𝕄, xᶠ𝕄)
     @. xᶜ𝕄 = -bᶜ𝕄
-    for ᶜ𝕋_name in filter(is_tracer_var, propertynames(xc))
+    for ᶜ𝕋_name in filter(CA.is_tracer_var, propertynames(xc))
         xᶜ𝕋 = getproperty(xc, ᶜ𝕋_name)
         bᶜ𝕋 = getproperty(bc, ᶜ𝕋_name)
         ∂ᶜ𝕋ₜ∂ᶠ𝕄 = getproperty(∂ᶜ𝕋ₜ∂ᶠ𝕄_field, ᶜ𝕋_name)
         @. xᶜ𝕋 = -bᶜ𝕋 + dtγ * apply(∂ᶜ𝕋ₜ∂ᶠ𝕄, xᶠ𝕄)
     end
-    for var_name in filter(is_edmf_var, propertynames(xc))
+    for var_name in filter(CA.is_edmf_var, propertynames(xc))
         xᶜ𝕋 = getproperty(xc, var_name)
         bᶜ𝕋 = getproperty(bc, var_name)
         @. xᶜ𝕋 = -bᶜ𝕋
     end
-    for var_name in filter(is_edmf_var, propertynames(xf))
+    for var_name in filter(CA.is_edmf_var, propertynames(xf))
         xᶠ𝕋 = getproperty(xf, var_name)
         bᶠ𝕋 = getproperty(bf, var_name)
         @. xᶠ𝕋 = -bᶠ𝕋
